+++ docs/project/changelog.md
@@ -19,6 +19,9 @@ myst:
 - {{ Enhancement }} Added built in packages for urllib3 and requests.
   {pr}`4332`
 
+- {{ Fix }} Fixed a bug that micropip would fail to install packages from pyodide-lock.json
+  if the package's name differs from its normalized name.
+  {pr}`4319`
 
 - {{ Enhancement }} Added a return type `PackageData` to `pyodide.loadPackage`,
   containing the loaded package's metadata.